Crate multiaddr

Implementation of multiaddr in Rust.

Re-exports

pub use multihash;

Macros

multiaddr

Easy way for a user to create a Multiaddr.

Structs

Iter

Iterator over Multiaddr Protocols.

Multiaddr

Representation of a Multiaddr.

Onion3Addr

Represents an Onion v3 address

PeerId

Identifier of a peer of the network.

ProtoStackIter

Iterator over the string idtenfiers of the protocols (not addrs) in a multiaddr

Enums

Error

Error types

FromUrlErr

Error while parsing an URL.

Protocol

Protocol describes all possible multiaddress protocols.

Functions

from_url

Attempts to parse an URL into a multiaddress.

from_url_lossy

Attempts to parse an URL into a multiaddress. Ignores possible loss of information.

Type Aliases

Result