Enum multiaddr::protocol::Protocol

pub enum Protocol {
    IP4,
    TCP,
    UDP,
    DCCP,
    IP6,
    SCTP,
    UTP,
    UDT,
    IPFS,
    HTTP,
    HTTPS,
    ONION,
}

Protocol

A type to describe the possible protocol used in a Multiaddr. Protocol is the list of all possible protocols.

Variants

IP4

IP4

TCP

TCP

UDP

UDP

DCCP

DCCP

IP6

IP6

SCTP

SCTP

UTP

UTP

UDT

UDT

IPFS

IPFS

HTTP

HTTP

HTTPS

HTTPS

ONION

Onion

Methods

impl Protocol

fn from_code(b: u16) -> Option<Protocol>

Convert a u16 based code to a Protocol.

Examples

use multiaddr::Protocol;

assert_eq!(Protocol::from_code(6u16), Some(Protocol::TCP));
assert_eq!(Protocol::from_code(455u16), None);

Failures

If no matching code is found None is returned.

fn size(&self) -> isize

Get the size from a Protocol.

Examples

use multiaddr::Protocol;

assert_eq!(Protocol::TCP.size(), 16);

fn from_name(s: &str) -> Option<Protocol>

Get the Protocol from a &str name.

Examples

use multiaddr::Protocol;

assert_eq!(Protocol::from_name("tcp").unwrap(), Protocol::TCP);

Failures

If no matching protocol is found None is returned.

fn address_string_to_bytes(&self, a: &str) -> Option<Vec<u8>>

Convert an array slice to the string representation.

Examples

use multiaddr::Protocol;

let proto = Protocol::IP4;
assert_eq!(proto.address_string_to_bytes("127.0.0.1").unwrap(), [127, 0, 0, 1]);

Failures

If there is no address representation for the protocol, like for https then None is returned.

fn bytes_to_string(&self, b: &[u8]) -> Option<String>

Convert an array slice to the string representation.

Examples

use multiaddr::Protocol;

let proto = Protocol::IP4;
let bytes = [127, 0, 0, 1];
assert_eq!(proto.bytes_to_string(&bytes).unwrap(), "127.0.0.1");

Failures

If there is no address representation for the protocol, like for https then None is returned.