Struct multiaddr::Multiaddr

pub struct Multiaddr {
    // some fields omitted
}

A multiaddr compatible network address.

Methods

impl Multiaddr

fn new() -> Multiaddr

Creates an empty Multiaddr.

fn from_parts(parts: Vec<Addr>) -> Multiaddr

Creates a Multiaddr from a series of address parts.

fn parts(&self) -> &Vec<Addr>

Gets an ordered list of address parts.

Examples

use multiaddr::{Multiaddr, Addr};

let ma = "/ip6/2001:db8:85a3::39/tcp/80".parse::<Multiaddr>().unwrap();
assert_eq!(ma.parts().len(), 2);
assert_eq!(ma.parts()[1], Addr::Tcp(80));

fn encapsulate(&self, other: &Multiaddr) -> Multiaddr

Wraps other.

Examples

use multiaddr::Multiaddr;

let foo = "/ip4/127.0.0.1/tcp/1723".parse::<Multiaddr>().unwrap();
let bar = "/ip4/1.2.3.4/udp/30".parse::<Multiaddr>().unwrap();
let proxied = foo.encapsulate(&bar);
assert_eq!(proxied.to_string(),
           "/ip4/127.0.0.1/tcp/1723/ip4/1.2.3.4/udp/30");

fn encapsulate_mut(&mut self, other: &Multiaddr)

Wraps other, in place.

fn write<T: Write>(&self, w: &mut T) -> Result<(), Error>

Writes a multiaddr in the standard binary encoding scheme.

fn to_bytes(&self) -> Result<Vec<u8>, Error>

Encodes a multiaddr to the standard binary encoding scheme.

fn from_bytes(buf: &[u8]) -> Result<Multiaddr, Error>

Decodes a multiaddr encoded in the standard binary encoding scheme.

fn is_empty(&self) -> bool

Returns true if the multiaddr doesn't contain any address parts.

fn contains_ipv4(&self) -> bool

Returns true if the multiaddr contains at least one IPv4 part.

fn contains_ipv6(&self) -> bool

Returns true if the multiaddr contains at least one IPv6 part.

fn contains_tcp(&self) -> bool

Returns true if the multiaddr contains at least one TCP part.

fn contains_udp(&self) -> bool

Returns true if the multiaddr contains at least one UDP part.

fn contains_dccp(&self) -> bool

Returns true if the multiaddr contains at least one DCCP part.

fn contains_ipfs(&self) -> bool

Returns true if the multiaddr contains at least one IPFS part.