Struct multiaddr::Multiaddr

pub struct Multiaddr { /* fields omitted */ }

Representation of a Multiaddr.

Methods

impl Multiaddr

fn new(input: &str) -> Result<Multiaddr>

Create a new multiaddr based on a string representation, like /ip4/127.0.0.1/udp/1234.

Examples

Simple construction

use multiaddr::Multiaddr;

let address = Multiaddr::new("/ip4/127.0.0.1/udp/1234").unwrap();
assert_eq!(address.to_bytes(), [
    4, 127, 0, 0, 1,
    17, 4, 210
]);

fn to_bytes(&self) -> Vec<u8>

Return a copy to disallow changing the bytes directly

fn as_slice(&self) -> &[u8]

Extracts a slice containing the entire underlying vector.

fn protocol(&self) -> Vec<Protocol>

Return a list of protocols

Examples

A single protocol

use multiaddr::{Multiaddr, Protocol};

let address = Multiaddr::new("/ip4/127.0.0.1").unwrap();
assert_eq!(address.protocol(), vec![Protocol::IP4]);

fn encapsulate<T: ToMultiaddr>(&self, input: T) -> Result<Multiaddr>

Wrap a given Multiaddr and return the combination.

Examples

use multiaddr::Multiaddr;

let address = Multiaddr::new("/ip4/127.0.0.1").unwrap();
let nested = address.encapsulate("/udt").unwrap();
assert_eq!(nested, Multiaddr::new("/ip4/127.0.0.1/udt").unwrap());

fn decapsulate<T: ToMultiaddr>(&self, input: T) -> Result<Multiaddr>

Remove the outer most address from itself.

Examples

use multiaddr::{Multiaddr, ToMultiaddr};

let address = Multiaddr::new("/ip4/127.0.0.1/udt/sctp/5678").unwrap();
let unwrapped = address.decapsulate("/udt").unwrap();
assert_eq!(unwrapped, Multiaddr::new("/ip4/127.0.0.1").unwrap());

assert_eq!(
    address.decapsulate("/udt").unwrap(),
    "/ip4/127.0.0.1/".to_multiaddr().unwrap()
);

Returns the original if the passed in address is not found

use multiaddr::ToMultiaddr;

let address = "/ip4/127.0.0.1/udt/sctp/5678".to_multiaddr().unwrap();
let unwrapped = address.decapsulate("/ip4/127.0.1.1").unwrap();
assert_eq!(unwrapped, address);